fix: setPaymentTransactionType covers giftcards and spacing issue fixed when saving sale methods
