+++ src/cartridges/int_adyen_overlay/cartridge/scripts/util/adyenConfigs.js
@@ -105,10 +105,6 @@ const adyenConfigsObj = {
     return getCustomPreference('AdyenTokenisationEnabled');
   },
 
-  getAdyenSalePaymentMethods: function () {
-    return getCustomPreference('AdyenSalePaymentMethods') ? getCustomPreference('AdyenSalePaymentMethods').toString().split(',') : '';
-  },
-
   getAdyenBasketFieldsEnabled() {
     return getCustomPreference('AdyenBasketFieldsEnabled');
   },
@@ -127,7 +123,7 @@ const adyenConfigsObj = {
 
   getAdyenSalePaymentMethods: function () {
     return getCustomPreference('AdyenSalePaymentMethods')
-      ? getCustomPreference('AdyenSalePaymentMethods').toString().split(',')
+      ? getCustomPreference('AdyenSalePaymentMethods').replace(/\s/g, '').toString().split(',')
       : [];
   },
